+++ hc/front/schemas.py
@@ -17,6 +17,7 @@ telegram_callback = {
                     "required": ["id", "type"],
                 },
                 "text": {"type": "string"},
+                "message_thread_id": {"type": "number"},
             },
             "required": ["chat", "text"],
         }